# app/app/openapi.yaml
...
openapi: 3.0.3
info:
    title: App API
    version: 0.0.1
paths:
    /api/admin_dhb/location_list:
        get:
            tags:
                - App
            operationId: App_AdminLocationList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminLocationListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/admin_dhb/reward_list:
        get:
            tags:
                - App
            operationId: App_AdminRewardList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminRewardListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/admin_dhb/user_list:
        get:
            tags:
                - App
            operationId: App_AdminUserList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminUserListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/admin_dhb/withdraw:
        get:
            tags:
                - App
            operationId: App_AdminWithdrawEth
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminWithdrawEthReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - App
            operationId: App_AdminWithdraw
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AdminWithdrawRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminWithdrawReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/admin_dhb/withdraw_list:
        get:
            tags:
                - App
            operationId: App_AdminWithdrawList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminWithdrawListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/deposit:
        get:
            tags:
                - App
            operationId: App_Deposit
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DepositReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/eth_authorize:
        post:
            tags:
                - App
            operationId: App_EthAuthorize
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EthAuthorizeRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EthAuthorizeReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/fee_reward_list:
        get:
            tags:
                - App
            operationId: App_FeeRewardList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FeeRewardListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/recommend_reward_list:
        get:
            tags:
                - App
            operationId: App_RecommendRewardList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RecommendRewardListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/reward_list:
        get:
            tags:
                - App
            operationId: App_RewardList
            parameters:
                - name: type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RewardListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/user_info:
        get:
            tags:
                - App
            operationId: App_UserInfo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserInfoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/withdraw:
        post:
            tags:
                - App
            operationId: App_Withdraw
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/WithdrawRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WithdrawReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/withdraw_list:
        get:
            tags:
                - App
            operationId: App_WithdrawList
            parameters:
                - name: type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WithdrawListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        AdminLocationListReply:
            type: object
            properties:
                locations:
                    type: array
                    items:
                        $ref: '#/components/schemas/AdminLocationListReply_LocationList'
        AdminLocationListReply_LocationList:
            type: object
            properties:
                createdAt:
                    type: string
                address:
                    type: string
                row:
                    type: integer
                    format: int64
                col:
                    type: integer
                    format: int64
                status:
                    type: string
                currentLevel:
                    type: integer
                    format: int64
                current:
                    type: string
                currentMax:
                    type: string
        AdminRewardListReply:
            type: object
            properties:
                rewards:
                    type: array
                    items:
                        $ref: '#/components/schemas/AdminRewardListReply_List'
...
        AdminUserListReply:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/AdminUserListReply_UserList'
...
        AdminWithdrawEthReply:
            type: object
            properties: {}
        AdminWithdrawListReply:
            type: object
            properties:
                withdraw:
                    type: array
                    items:
                        $ref: '#/components/schemas/AdminWithdrawListReply_List'
...
        AdminWithdrawReply:
            type: object
            properties:
                status:
                    type: string
        AdminWithdrawRequest_SendBody:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
        DepositReply:
            type: object
            properties: {}
        EthAuthorizeReply:
            type: object
            properties:
                token:
                    type: string
        EthAuthorizeRequest_SendBody:
            type: object
            properties:
                address:
                    type: string
                code:
                    type: string
        FeeRewardListReply:
            type: object
            properties:
                rewards:
                    type: array
                    items:
                        $ref: '#/components/schemas/FeeRewardListReply_List'
...
        RecommendRewardListReply:
            type: object
            properties:
                rewards:
                    type: array
                    items:
                        $ref: '#/components/schemas/RecommendRewardListReply_List'
...
        RewardListReply:
            type: object
            properties:
                rewards:
                    type: array
                    items:
                        $ref: '#/components/schemas/RewardListReply_List'
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        WithdrawListReply:
            type: object
            properties:
                withdraw:
                    type: array
                    items:
                        $ref: '#/components/schemas/WithdrawListReply_List'
...
        WithdrawReply:
            type: object
            properties:
                status:
                    type: string
        WithdrawRequest_SendBody:
            type: object
            properties:
                type:
                    type: string
                amount:
                    type: string
...